fix(diet): 修复饮食记录保存失败问题
- 修复calories类型不匹配(int -> float)与数据库模型一致
- 后端API添加详细错误日志和异常处理
- 前端改进错误处理和调试日志输出
- 修复image_type为null时的处理逻辑

+++ backend/app/schemas/diet.py
@@ -19,7 +19,7 @@ class DietRecordBase(BaseModel):
     meal_type: MealType
     meal_time: Optional[time] = None
     food_items: str  # 食物列表，逗号分隔
-    calories: Optional[int] = None  # 热量 (kcal)
+    calories: Optional[float] = None  # 热量 (kcal) - 使用float与数据库一致
     protein: Optional[float] = None  # 蛋白质 (g)
     carbs: Optional[float] = None  # 碳水化合物 (g)
     fat: Optional[float] = None  # 脂肪 (g)
@@ -42,7 +42,7 @@ class DietRecordUpdate(BaseModel):
     meal_type: Optional[MealType] = None
     meal_time: Optional[time] = None
     food_items: Optional[str] = None
-    calories: Optional[int] = None
+    calories: Optional[float] = None  # 使用float与数据库一致
     protein: Optional[float] = None
     carbs: Optional[float] = None
     fat: Optional[float] = None
